dtl: more cleanup
